module rewrite: refactor rollback for npm mods
