+++ lib/API/Modules/LOCAL.js
@@ -10,8 +10,6 @@ var Configuration = require('../../Configuration.js');
 var cst           = require('../../../constants.js');
 var Common        = require('../../Common');
 var Utility       = require('../../Utility.js');
-var Rollback = require('./Rollback.js')
-var StartModule = require('./StartModule.js')
 var readline = require('readline')
 
 var INTERNAL_MODULES = {
