+++ lib/API/Modules/NPM.js
@@ -8,7 +8,6 @@ var Configuration = require('../../Configuration.js');
 var cst           = require('../../../constants.js');
 var Common        = require('../../Common');
 var Utility       = require('../../Utility.js');
-var Rollback = require('./Rollback.js')
 var readline = require('readline')
 
 module.exports = {
@@ -152,8 +151,9 @@ function publish(opts, cb) {
 function moduleExistInLocalDB(CLI, module_name, cb) {
   var modules = Configuration.getSync(cst.MODULE_CONF_PREFIX);
   if (!modules) return cb(false);
+  var module_name_only = Utility.getCanonicModuleName(module_name)
   modules = Object.keys(modules);
-  return cb(modules.indexOf(module_name) > -1 ? true : false);
+  return cb(modules.indexOf(module_name_only) > -1 ? true : false);
 };
 
 function install(CLI, module_name, opts, cb) {
@@ -276,11 +276,11 @@ function start(PM2, modules, module_name, cb) {
 }
 
 function uninstall(CLI, module_name, cb) {
-  var proc_path = path.join(cst.DEFAULT_MODULE_PATH, module_name);
+  var module_name_only = Utility.getCanonicModuleName(module_name)
+  var proc_path = path.join(cst.DEFAULT_MODULE_PATH, module_name_only);
+  Configuration.unsetSync(cst.MODULE_CONF_PREFIX + ':' + module_name_only);
 
-  Configuration.unsetSync(cst.MODULE_CONF_PREFIX + ':' + module_name);
-
-  CLI.deleteModule(module_name, function(err, data) {
+  CLI.deleteModule(module_name_only, function(err, data) {
     if (err) {
       Common.printError(err);
 
@@ -360,7 +360,7 @@ function StartModule(CLI, opts, cb) {
       return setTimeout(function() {
         CLI.describe(package_json.name, function(err, apps) {
           if (err || apps[0].pm2_env.restart_time > 2) {
-            return require('./Rollback.js').revert(CLI, package_json.name, function() {
+            return Rollback.revert(CLI, package_json.name, function() {
               return cb(new Error('New Module is instable, restored to previous version'));
             });
           }
@@ -372,3 +372,46 @@ function StartModule(CLI, opts, cb) {
     return cb(null, data);
   });
 };
+
+
+
+var Rollback = {
+  revert : function(CLI, module_name, cb) {
+    var canonic_module_name = Utility.getCanonicModuleName(module_name);
+    var backup_path = path.join(require('os').tmpdir(), canonic_module_name);
+    var module_path = path.join(cst.DEFAULT_MODULE_PATH, canonic_module_name);
+
+    try {
+      fs.statSync(backup_path)
+    } catch(e) {
+      return cb(new Error('no backup found'));
+    }
+
+    Common.printOut(cst.PREFIX_MSG_MOD + chalk.bold.red('[[[[[ Module installation failure! ]]]]]'));
+    Common.printOut(cst.PREFIX_MSG_MOD + chalk.bold.red('[RESTORING TO PREVIOUS VERSION]'));
+
+    CLI.deleteModule(canonic_module_name, function() {
+      // Delete failing module
+      require('shelljs').rm('-r', module_path);
+      // Restore working version
+      require('shelljs').cp('-r', backup_path, cst.DEFAULT_MODULE_PATH);
+
+      var proc_path = path.join(module_path, 'node_modules', canonic_module_name);
+      var package_json_path = path.join(proc_path, 'package.json');
+
+      // Start module
+      StartModule(CLI, {
+        cmd : package_json_path,
+        development_mode : false,
+        proc_path : proc_path
+      }, cb);
+    });
+  },
+  backup : function(module_name) {
+    // Backup current module
+    var tmpdir = require('os').tmpdir();
+    var canonic_module_name = Utility.getCanonicModuleName(module_name);
+    var module_path = path.join(cst.DEFAULT_MODULE_PATH, canonic_module_name);
+    require('shelljs').cp('-r', module_path, tmpdir);
+  }
+}
