--- lib/API/Modules/StartModule.js
@@ -1,15 +1,0 @@
-var path          = require('path');
-var fs            = require('fs');
-var os            = require('os');
-var parallel      = require('async/parallel');
-var eachLimit     = require('async/eachLimit');
-var forEachLimit  = require('async/forEachLimit');
-var p             = path;
-var readline      = require('readline');
-var spawn         = require('child_process').spawn;
-var chalk         = require('chalk');
-var Configuration = require('../../Configuration.js');
-var cst           = require('../../../constants.js');
-var Common        = require('../../Common');
-var Utility       = require('../../Utility.js');
-var Rollback = require('./Rollback.js')
