+++ lib/API/Modules/TAR.js
@@ -2,8 +2,6 @@
 var Configuration = require('../../Configuration.js');
 var cst = require('../../../constants.js');
 var Common = require('../../Common');
-var Rollback = require('./Rollback.js')
-var StartModule = require('./StartModule.js')
 var forEachLimit  = require('async/forEachLimit');
 
 var path = require('path');
@@ -173,7 +171,12 @@ function getModuleName(module_filepath, cb) {
 }
 
 function publish(opts, cb) {
-  var pkg = require(path.join(process.cwd(), 'package.json'))
+  try {
+    var pkg = require(path.join(process.cwd(), 'package.json'))
+  } catch(e) {
+    Common.err(`${process.cwd()} module does not contain any package.json`)
+    process.exit(1)
+  }
 
   if (!pkg.name) throw new Error('Attribute name should be present')
   if (!pkg.version) throw new Error('Attribute version should be present')
